Refactor BoardGroup into a separate file.
This is to prevent circular dependency issues
