+++ j5/boards/__init__.py
@@ -1,5 +1,6 @@
 """This module contains the boards that we support."""
 
-from .board import Board, BoardGroup
+from .board import Board
+from .board_group import BoardGroup
 
 __all__ = ["Board", "BoardGroup"]
